fixes on Drashna's sugg
