+++ keyboards/h0oni/deskpad/deskpad.h
@@ -24,3 +24,5 @@
     { K00,  K01,  K02 },\
     { K10,  K11,  K12 } \
 }
+
+
