+++ keyboards/h0oni/deskpad/keymaps/default/rules.mk
@@ -1,0 +1,1 @@
+TAP_DANCE_ENABLE = yes
