+++ keyboards/h0oni/deskpad/keymaps/via/rules.mk
@@ -1,2 +1,3 @@
 VIA_ENABLE = yes
 LTO_ENABLE = yes
+TAP_DANCE_ENABLE = yes
