+++ keyboards/h0oni/deskpad/rules.mk
@@ -20,4 +20,3 @@ BACKLIGHT_ENABLE = no       # Enable keyboard backlight functionality
 RGBLIGHT_ENABLE = no       # Enable keyboard RGB underglow
 AUDIO_ENABLE = no           # Audio output
 ENCODER_ENABLE = yes
-TAP_DANCE_ENABLE = yes
